requirements.txt
appdirs==1.4.4
asgiref==3.4.1
backports.zoneinfo==0.2.1
black==21.7b0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.0.1
coreapi==2.3.3
coreschema==0.0.4
datetime-truncate==1.1.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.6
django-admin-charts==1.2.3
django-admin-volt==1.0.4
django-bower==5.2.0
django-cors-headers==3.7.0
django-import-export==3.0.1
django-jazzmin==2.6.0
django-memoize==2.3.1
django-multiselectfield==0.1.12
django-nvd3==0.9.7
django-rest-swagger==2.2.0
djangorestframework==3.12.4
et-xmlfile==1.1.0
gunicorn==20.1.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupPy==1.14
MarkupSafe==2.1.1
mypy-extensions==0.4.3
odfpy==1.4.1
openapi-codec==1.3.2
openpyxl==3.0.10
pathspec==0.9.0
Pillow==8.3.1
psycopg2-binary==2.9.1
python-dateutil==2.8.2
python-nvd3==0.14.2
python-slugify==1.1.4
pytz==2021.1
PyYAML==6.0
regex==2021.8.3
requests==2.28.1
simplejson==3.18.0
six==1.16.0
sqlparse==0.4.1
tablib==3.2.1
tomli==1.2.0
Unidecode==1.3.6
uritemplate==4.1.1
urllib3==1.26.12
whitenoise==5.3.0
xlrd==2.0.1
xlwt==1.3.0
